# docs/openapi/orders.yaml
openapi: 3.0.3
info:
  title: orders-service API
  description: |
    Microservicio de gestión de pedidos con trazabilidad completa y métricas de eficiencia.

    **Funcionalidades principales:**
    - Gestión completa del ciclo de vida de pedidos
    - Trazabilidad en tiempo real con MongoDB
    - Métricas de eficiencia de restaurantes y empleados (HU-018)
    - Control de acceso basado en roles (JWT)

    **Arquitectura:**
    - Hexagonal Architecture
    - PostgreSQL para datos transaccionales
    - MongoDB para eventos de trazabilidad
    - Integración con Restaurant Service para validación de autorización
  version: 1.0.0
  contact:
    name: Orders Service Team
    email: orders@pragma.com.co
...
servers:
  - url: http://localhost:8082
    description: orders-service local development
  - url: https://orders-api.pragma.com.co
    description: orders-service production
security:
  - bearerAuth: []
paths:
...
  /api/v1/orders/{id}/traceability:
    get:
      tags: [Traceability]
      summary: Get order traceability events (ALL AUTHENTICATED ROLES)
      description: |
        Returns traceability events for a specific order with complete state change history.

        **Authorization rules:**
        - **CUSTOMER**: Can only view traceability of their own orders
        - **EMPLOYEE**: Can view traceability of orders from their assigned restaurant
        - **OWNER**: Can view traceability of orders from their restaurant
        - **ADMIN**: Can view traceability of any order

        **Response includes:**
        - Complete state transition history (PENDIENTE → EN_PREPARACION → LISTO → ENTREGADO)
        - Timestamps for each state change
        - Employee information who performed each action (when applicable)
        - MongoDB-based storage for scalable traceability
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: OK - Complete traceability history returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TraceabilityEventResponse'
              example:
                - id: "68b70c41c8b0271f40963e4b"
                  orderId: 12
                  eventType: null
                  previousStatus: null
                  newStatus: "PENDIENTE"
                  timestamp: "2025-09-02T09:30:46.291"
                  employeeId: null
                  employeeName: null
                - id: "68b70c5ec8b0271f40963e4c"
                  orderId: 12
                  eventType: null
                  previousStatus: "PENDIENTE"
                  newStatus: "EN_PREPARACION"
                  timestamp: "2025-09-02T09:31:24.299"
                  employeeId: 3
                  employeeName: "employee"
                - id: "68b70c76c8b0271f40963e4d"
                  orderId: 12
                  eventType: null
                  previousStatus: "EN_PREPARACION"
                  newStatus: "LISTO"
                  timestamp: "2025-09-02T09:31:48.821"
                  employeeId: 3
                  employeeName: "employee"
                - id: "68b70c8ec8b0271f40963e4e"
                  orderId: 12
                  eventType: null
                  previousStatus: "LISTO"
                  newStatus: "ENTREGADO"
                  timestamp: "2025-09-02T09:32:14.542"
                  employeeId: 3
                  employeeName: "employee"
        '401': { description: Unauthorized - Missing or invalid JWT token }
        '403': { description: Forbidden - You are not authorized to view this order's traceability }
        '404': { description: Order not found }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtenido del servicio de autenticación.

        **Formato:** `Authorization: Bearer <token>`

        **Roles soportados:**
        - `CUSTOMER`: Clientes del sistema
        - `EMPLOYEE`: Empleados de restaurantes
        - `OWNER`: Propietarios de restaurantes
        - `ADMIN`: Administradores del sistema
  schemas:
...
    TraceabilityEventResponse:
      type: object
      description: |
        Represents a single traceability event in an order's lifecycle.
        Each event captures a state transition with complete context information.
      properties:
        id:
          type: string
          description: Unique MongoDB ObjectId for the traceability event
          example: "68b70c41c8b0271f40963e4b"
        orderId:
          type: integer
          format: int64
          description: ID of the order this event belongs to
          example: 12
        eventType:
          type: string
          enum: [ORDER_CREATED, STATUS_CHANGED]
          nullable: true
          description: Type of event (currently not used, reserved for future expansion)
        previousStatus:
          type: string
          enum: [PENDIENTE, EN_PREPARACION, LISTO, ENTREGADO, CANCELADO]
          nullable: true
          description: Previous order status (null for initial order creation)
          example: "PENDIENTE"
        newStatus:
          type: string
          enum: [PENDIENTE, EN_PREPARACION, LISTO, ENTREGADO, CANCELADO]
          description: New order status after the state transition
          example: "EN_PREPARACION"
        timestamp:
          type: string
          format: date-time
          description: Exact timestamp when the state change occurred
          example: "2025-09-02T09:31:24.299"
        employeeId:
          type: integer
          format: int64
          nullable: true
          description: ID of the employee who performed the action (null for system-generated events)
          example: 3
        employeeName:
          type: string
          nullable: true
          description: |
            Name of the employee who performed the action.
            Resolved from Users service with fallback logic:
            1. firstName + lastName (if available)
            2. name + surname (alternative mapping)
            3. email username (as fallback)
            null for system-generated events like order creation.
          example: "employee"
      required:
        - id
        - orderId
        - newStatus
        - timestamp
...